add list of manually entered IDs
This is a list of project IDs that have to be added to the archive manually, likely because they are private and can only be accessed via a link that has been freely shared by the user who created the project or someone who it was shared with. More IDs will be added for archival as more are freely shared with others.